chore: move LocaleSwitcher to $lib/components
