+++ src/lib/components/LocaleSwitcher/LocaleSwitcher.svelte
